・エリア基底クラスを定義。 ・Areaクラスの機能を基底クラス側に移動。
