+++ Server/AnpanMMOServer/AnpanMMOServer.vcxproj
@@ -126,6 +126,7 @@
   <ItemGroup>
     <ClInclude Include="AnpanManager.h" />
     <ClInclude Include="AnpanPopArea.h" />
+    <ClInclude Include="Area\AreaBase.h" />
     <ClInclude Include="Area\AreaConfig.h" />
     <ClInclude Include="Area\Area.h" />
     <ClInclude Include="Area\AreaManager.h" />
@@ -211,6 +212,7 @@
   <ItemGroup>
     <ClCompile Include="AnpanManager.cpp" />
     <ClCompile Include="AnpanPopArea.cpp" />
+    <ClCompile Include="Area\AreaBase.cpp" />
     <ClCompile Include="Area\AreaConfig.cpp" />
     <ClCompile Include="Area\Area.cpp" />
     <ClCompile Include="Area\AreaManager.cpp" />
